Move noseclasses into testing.__init__
